chore: move entry w/ releaseId tests to existing entry integration tests file
